refactor(pedidos): improve module quality without functional changes
T1 — Fix type re-export chain:
  - index.ts now exports types from ./types (canonical source)
  - Removed redundant re-export passthrough in mock-data.ts
  - Updated 6 internal components: import types from ../types

T2 — Remove stale debug logging:
  - Removed console.info('DEBUG loadOrdersHydrated...') from order.loader.ts

T3 — Remove dead code (3 files, 324 lines):
  - Deleted order-filters.tsx (zero consumers)
  - Deleted order-item.tsx (zero consumers)
  - Deleted orders-list.tsx (zero consumers — old list view replaced by UniversalListView)

T4 — Reduce bulk action duplication:
  - Extracted executeBulkAction() shared helper
  - 3 public functions now delegate to helper (96 -> 77 lines)
  - Same behavior, same audit trail, same revalidation

Zero functional changes. No logic changes. No schema changes.

QG Results:
- typecheck: PASS
- lint:env: PASS
- lint:pii: PASS
- db:verify: PASS
- routes:inventory: PASS (342 routes)
- routes:verify: PASS (342 routes)
- routes:verify-security: PASS (160 routes)
- test:ci: PASS (917 tests, 187 files)
- check:boundaries: PASS (1264 files)
- build: PASS
- release:check: RELEASE_CHECK_PASS
- npm audit: 0 vulnerabilities
