+++ src/modules/pedidos/actions/bulk.ts
@@ -7,12 +7,22 @@ import { getTenantId } from '@/modules/audit/audit.actions';
 import { operationalAuditService } from '@/modules/audit/operational-audit.service';
 import { revalidatePath } from 'next/cache';
 
-export async function bulkApproveOrders(orderIds: string[]) {
+// ── Shared helper ──────────────────────────────────────────────────────────
+
+interface BulkActionParams {
+    orderIds: string[];
+    setValues: Record<string, unknown>;
+    actionType: string;
+    auditState: Record<string, unknown>;
+}
+
+async function executeBulkAction({ orderIds, setValues, actionType, auditState }: BulkActionParams) {
     if (!orderIds.length) return { success: false };
+
     const tenantId = await getTenantId();
-    
+
     await db.update(orders)
-        .set({ status: 'processando', updatedAt: new Date() })
+        .set({ ...setValues, updatedAt: new Date() })
         .where(
             and(
                 eq(orders.tenantId, tenantId),
@@ -25,8 +35,8 @@ export async function bulkApproveOrders(orderIds: string[]) {
             tenantId,
             entityType: 'order',
             entityId: id,
-            actionType: 'bulk_order_approved',
-            afterState: { status: 'processando' },
+            actionType,
+            afterState: auditState,
             origin: 'bulk',
             actorId: 'Sistema'
         });
@@ -34,62 +44,33 @@ export async function bulkApproveOrders(orderIds: string[]) {
 
     revalidatePath('/pedidos');
     return { success: true };
+}
+
+// ── Public actions ─────────────────────────────────────────────────────────
+
+export async function bulkApproveOrders(orderIds: string[]) {
+    return executeBulkAction({
+        orderIds,
+        setValues: { status: 'processando' },
+        actionType: 'bulk_order_approved',
+        auditState: { status: 'processando' },
+    });
 }
 
 export async function bulkAssignOrderOwner(orderIds: string[], ownerId: string) {
-    if (!orderIds.length) return { success: false };
-    const tenantId = await getTenantId();
-    
-    await db.update(orders)
-        .set({ ownerId: ownerId, updatedAt: new Date() })
-        .where(
-            and(
-                eq(orders.tenantId, tenantId),
-                inArray(orders.id, orderIds)
-            )
-        );
-
-    for (const id of orderIds) {
-        await operationalAuditService.logActivity({
-            tenantId,
-            entityType: 'order',
-            entityId: id,
-            actionType: 'bulk_order_owner_assigned',
-            afterState: { ownerId },
-            origin: 'bulk',
-            actorId: 'Sistema'
-        });
-    }
-
-    revalidatePath('/pedidos');
-    return { success: true };
+    return executeBulkAction({
+        orderIds,
+        setValues: { ownerId },
+        actionType: 'bulk_order_owner_assigned',
+        auditState: { ownerId },
+    });
 }
 
 export async function bulkCancelOrders(orderIds: string[]) {
-    if (!orderIds.length) return { success: false };
-    const tenantId = await getTenantId();
-    
-    await db.update(orders)
-        .set({ status: 'cancelado', updatedAt: new Date() })
-        .where(
-            and(
-                eq(orders.tenantId, tenantId),
-                inArray(orders.id, orderIds)
-            )
-        );
-
-    for (const id of orderIds) {
-        await operationalAuditService.logActivity({
-            tenantId,
-            entityType: 'order',
-            entityId: id,
-            actionType: 'bulk_order_cancelled',
-            afterState: { status: 'cancelado' },
-            origin: 'bulk',
-            actorId: 'Sistema'
-        });
-    }
-
-    revalidatePath('/pedidos');
-    return { success: true };
+    return executeBulkAction({
+        orderIds,
+        setValues: { status: 'cancelado' },
+        actionType: 'bulk_order_cancelled',
+        auditState: { status: 'cancelado' },
+    });
 }
